+++ Runtime/Extensions/StringExtensions.cs
@@ -62,21 +62,21 @@ namespace Mane.Extensions
         public static string FormatColored(this string str, string format, float value, Color color) =>
             string.Format(str, $"<color={color.ToHex()}>{value.ToString(format)}</color>");
 
-        public static string FormatColored(this string str, string format, 
-            float v1, Color c1,
-            float v2, Color c2) =>
+        public static string FormatColored(this string str, 
+            string f1, float v1, Color c1,
+            string f2, float v2, Color c2) =>
             string.Format(str,
-                $"<color={c1.ToHex()}>{v1.ToString(format)}</color>",
-                $"<color={c2.ToHex()}>{v2.ToString(format)}</color>");
+                $"<color={c1.ToHex()}>{v1.ToString(f1)}</color>",
+                $"<color={c2.ToHex()}>{v2.ToString(f2)}</color>");
 
-        public static string FormatColored(this string str, string format, 
-            float v1, Color c1,
-            float v2, Color c2,
-            float v3, Color c3) =>
+        public static string FormatColored(this string str,
+            string f1, float v1, Color c1,
+            string f2, float v2, Color c2,
+            string f3, float v3, Color c3) =>
             string.Format(str,
-                $"<color={c1.ToHex()}>{v1.ToString(format)}</color>",
-                $"<color={c2.ToHex()}>{v2.ToString(format)}</color>",
-                $"<color={c3.ToHex()}>{v3.ToString(format)}</color>");
+                $"<color={c1.ToHex()}>{v1.ToString(f1)}</color>",
+                $"<color={c2.ToHex()}>{v2.ToString(f2)}</color>",
+                $"<color={c3.ToHex()}>{v3.ToString(f3)}</color>");
         
         public static string FormatColored(this string str,
             params (float value, string format, Color color)[] args)