+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.6.24:
+    - Added PrefabList<T> to easy UI list managing.
+    - Fixed some formatting issues in string.FormColored().
+
 1.6.23:
     - Added string.FormatColored() extensions that duplicate base Format method, but with rich text coloring for float and int values.